Localtime reset for mergeList self case

+++ distMembership/src/main/java/edu/uiuc/cs425/MembershipListStruct.java
@@ -17,7 +17,7 @@ public class MembershipListStruct {
 		m_bSuspect = false;
 	}
 	
-	public void ResetLocalTime(int localTime)
+	public void ResetLocalTime(long localTime)
 	{
 		m_nLocalTime = localTime;
 	}
